Revert "1"
This reverts commit 9e7ae2293f316de22ff3cc7b08066277446ae926.

+++ src/main/java/ru/yandex/practicum/sleeptracker/analysis/Avg.java
@@ -11,7 +11,7 @@ public class Avg implements Function<List<SleepingSession>, SleepAnalysisResult>
     @Override
     public SleepAnalysisResult apply(List<SleepingSession> o) {
         return new SleepAnalysisResult(
-                Math.round(
+                (int) Math.round(
                     o.stream()
                         .map(session -> (int) Duration.between(session.begin, session.end).toMinutes())
                         .mapToInt(i -> i)
